refactor: Split MessageEvent into message_event.go
